Update ennemie stats

+++ Assets/Scripts/BoatData.cs
@@ -2,12 +2,12 @@ using UnityEngine;
 
 public class BoatData : MonoBehaviour
 {
-    public int boatMaxPv = 100;
-    public int boatPv = 100;
+    public float boatMaxPv = 100;
+    public float boatPv = 100;
 
     public int boatDamage = 10;
 
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         boatPv -= dmg;
         if (boatPv <= 0)
